+++ packages/ti/osal/soc/j7200/TimerP_default_a72.c
@@ -52,147 +52,147 @@ TimerP_dmTimerDefault gDmTimerPInfoTbl[TimerP_numTimerDevices] = {
   /* Timer ID 0 */
   {
     "DMTimer0",
-    (uint32_t)CSL_TIMER0_CFG_BASE,
+    (uintptr_t)CSL_TIMER0_CFG_BASE,
     (int32_t)CSLR_COMPUTE_CLUSTER0_GIC500SS_SPI_TIMER0_INTR_PEND_0,
     TIMERP_EVENT_NOT_AVAILABLE
   },
   /* Timer ID 1 */
   {
     "DMTimer1",
-    (uint32_t)CSL_TIMER1_CFG_BASE,
+    (uintptr_t)CSL_TIMER1_CFG_BASE,
     (int32_t)CSLR_COMPUTE_CLUSTER0_GIC500SS_SPI_TIMER1_INTR_PEND_0,
     TIMERP_EVENT_NOT_AVAILABLE
   },
   /* Timer ID 2 */
   {
     "DMTimer2",
-    (uint32_t)CSL_TIMER2_CFG_BASE,
+    (uintptr_t)CSL_TIMER2_CFG_BASE,
     (int32_t)CSLR_COMPUTE_CLUSTER0_GIC500SS_SPI_TIMER2_INTR_PEND_0,
     TIMERP_EVENT_NOT_AVAILABLE
   },
   /* Timer ID 3 */
   {
     "DMTimer3",
-    (uint32_t)CSL_TIMER3_CFG_BASE,
+    (uintptr_t)CSL_TIMER3_CFG_BASE,
     (int32_t)CSLR_COMPUTE_CLUSTER0_GIC500SS_SPI_TIMER3_INTR_PEND_0,
     TIMERP_EVENT_NOT_AVAILABLE
   },
   /* Timer ID 4 */
   {
     "DMTimer4",
-    (uint32_t)CSL_TIMER4_CFG_BASE,
+    (uintptr_t)CSL_TIMER4_CFG_BASE,
     (int32_t)CSLR_COMPUTE_CLUSTER0_GIC500SS_SPI_TIMER4_INTR_PEND_0,
     TIMERP_EVENT_NOT_AVAILABLE
   },
   /* Timer ID 5 */
   {
     "DMTimer5",
-    (uint32_t)CSL_TIMER5_CFG_BASE,
+    (uintptr_t)CSL_TIMER5_CFG_BASE,
     (int32_t)CSLR_COMPUTE_CLUSTER0_GIC500SS_SPI_TIMER5_INTR_PEND_0,
     TIMERP_EVENT_NOT_AVAILABLE
   },
   /* Timer ID 6 */
   {
     "DMTimer6",
-    (uint32_t)CSL_TIMER6_CFG_BASE,
+    (uintptr_t)CSL_TIMER6_CFG_BASE,
     (int32_t)CSLR_COMPUTE_CLUSTER0_GIC500SS_SPI_TIMER6_INTR_PEND_0,
     TIMERP_EVENT_NOT_AVAILABLE
   },
    /* Timer ID 7 */
   {
     "DMTimer7",
-    (uint32_t)CSL_TIMER7_CFG_BASE,
+    (uintptr_t)CSL_TIMER7_CFG_BASE,
     (int32_t)CSLR_COMPUTE_CLUSTER0_GIC500SS_SPI_TIMER7_INTR_PEND_0,
     TIMERP_EVENT_NOT_AVAILABLE
   },
   /* Timer ID 8 */
   {
     "DMTimer8",
-    (uint32_t)0x0,
+    (uintptr_t)0x0,
     (int32_t)CSLR_MCU_R5FSS0_CORE0_INTR_MCU_TIMER8_INTR_PEND_0,
     TIMERP_EVENT_NOT_AVAILABLE
   },
   /* Timer ID 9 */
   {
     "DMTimer9",
-    (uint32_t)0x0,
+    (uintptr_t)0x0,
     (int32_t)CSLR_MCU_R5FSS0_CORE0_INTR_MCU_TIMER9_INTR_PEND_0,
     TIMERP_EVENT_NOT_AVAILABLE
   },
   /* Timer ID 10 */
   {
     "DMTimer10",
-    (uint32_t)0x0,
+    (uintptr_t)0x0,
     (int32_t)CSLR_COMPUTE_CLUSTER0_GIC500SS_SPI_TIMER10_INTR_PEND_0,
     TIMERP_EVENT_NOT_AVAILABLE
   },
   /* Timer ID 11 */
   {
     "DMTimer11",
-    (uint32_t)0x0,
+    (uintptr_t)0x0,
     (int32_t)CSLR_COMPUTE_CLUSTER0_GIC500SS_SPI_TIMER11_INTR_PEND_0,
     TIMERP_EVENT_NOT_AVAILABLE
   },
   /* Timer ID 12 */
   {
     "DMTimer12",
-    (uint32_t)0x0,
+    (uintptr_t)0x0,
     (int32_t)CSLR_COMPUTE_CLUSTER0_GIC500SS_SPI_TIMER12_INTR_PEND_0,
     TIMERP_EVENT_NOT_AVAILABLE
   },
   /* Timer ID 13 */
   {
     "DMTimer13",
-    (uint32_t)0x0,
+    (uintptr_t)0x0,
     (int32_t)CSLR_COMPUTE_CLUSTER0_GIC500SS_SPI_TIMER13_INTR_PEND_0,
     TIMERP_EVENT_NOT_AVAILABLE
   },
   /* Timer ID 14 */
   {
     "DMTimer14",
-    (uint32_t)0x0,
+    (uintptr_t)0x0,
     (int32_t)CSLR_COMPUTE_CLUSTER0_GIC500SS_SPI_TIMER14_INTR_PEND_0,
     TIMERP_EVENT_NOT_AVAILABLE
   },
   /* Timer ID 15 */
   {
     "DMTimer15",
-    (uint32_t)0x0,
+    (uintptr_t)0x0,
     (int32_t)CSLR_COMPUTE_CLUSTER0_GIC500SS_SPI_TIMER15_INTR_PEND_0,
     TIMERP_EVENT_NOT_AVAILABLE
   },
   /* Timer ID 16 */
   {
     "DMTimer16",
-    (uint32_t)0x0,
+    (uintptr_t)0x0,
     (int32_t)CSLR_COMPUTE_CLUSTER0_GIC500SS_SPI_TIMER16_INTR_PEND_0,
     TIMERP_EVENT_NOT_AVAILABLE
   },
   /* Timer ID 17 */
   {
     "DMTimer17",
-    (uint32_t)0x0,
+    (uintptr_t)0x0,
     (int32_t)CSLR_COMPUTE_CLUSTER0_GIC500SS_SPI_TIMER17_INTR_PEND_0,
     TIMERP_EVENT_NOT_AVAILABLE
   },
   /* Timer ID 18 */
   {
     "DMTimer18",
-    (uint32_t)0x0,
+    (uintptr_t)0x0,
     (int32_t)CSLR_COMPUTE_CLUSTER0_GIC500SS_SPI_TIMER18_INTR_PEND_0,
     TIMERP_EVENT_NOT_AVAILABLE
   },
   /* Timer ID 19 */
   {
     "DMTimer19",
-    (uint32_t)0x0,
+    (uintptr_t)0x0,
     (int32_t)CSLR_COMPUTE_CLUSTER0_GIC500SS_SPI_TIMER19_INTR_PEND_0,
     TIMERP_EVENT_NOT_AVAILABLE
   }
 };
 
 /* Returns the default frequency lower 32 bits */
-int32_t TimerP_getDefaultFreqLo(uint32_t timerId)
+uint32_t TimerP_getDefaultFreqLo(uint32_t timerId)
 {
     (void)timerId;
     /* return the same default frequency for all timer Ids */
@@ -200,7 +200,7 @@ int32_t TimerP_getDefaultFreqLo(uint32_t timerId)
 }
 
 /* Returns the default frequency higher 32 bits */
-int32_t TimerP_getDefaultFreqHi(uint32_t timerId)
+uint32_t TimerP_getDefaultFreqHi(uint32_t timerId)
 {
     (void)timerId;
     /* return the same default frequency for all timer Ids */
